Add files via upload
The project is being run via Xampp, and all the files should be downloaded inside a newly created folder in the htdocs directory. After this is complete activate the xampp database make sure to go to phpmyadmin to create the database with the proper columns and the project should be able to run.